feat: rename classify_domain* files, switch pipeline to fit-derivation classifier, refine system prompt
- Rename classify_domain.py -> classify_domain_v1.py (legacy single-label
  variant), classify_domain2.py -> classify_domain_v2.py,
  classify_domain3.py -> classify_domain_v3.py.
- Promote former classify_domain4.py to classify_domain.py as the
  current best implementation (fit-derivation: LLM emits only fits +
  confidence + rationale; primary/secondaries derived in code).
- Point worker_plan_internal/plan/nodes/classify_domain.py at the new
  classify_domain (was classify_domain_v1).
- System-prompt refinements driven by the qwen3-30b smoke run:
  - Strengthen Personal-vs-Software framing for prompts where a
    single individual uses an off-the-shelf app/AI for a personal
    goal (the app is role="tool", not the outcome).
  - Sharpen Education-vs-Hospitality contrast for community classes,
    drop-in workshops, and recurring courses run from a venue.
  - Add a role-enum reminder: role must be exactly one of the seven
    listed literals; this fixed an `'environment'` validation error
    seen on qwen.
- Smoke harness in __main__: progressive seed-exclusion logic so
  fresh samples never overlap with prompts already tested under
  seeds 7, 8, 100. Seed 300 sample size set to 20.
